+++ src/senzing/__init__.py
@@ -7,13 +7,14 @@
 
 # Step 1: Import the files so that the __all__ attribute will work with the "name" (e.g. G2Config, G2ConfigMgr)
 
-from . import G2Config, G2ConfigMgr, G2Diagnostic, G2Engine, G2Exception, G2Hasher, G2IniParams, G2Product
+from . import G2Config, G2ConfigMgr, G2Diagnostic, G2Engine, G2EngineFlags, G2Exception, G2Hasher, G2IniParams, G2Product
 
 import_lists = [
     G2Config.__all__,
     G2ConfigMgr.__all__,
     G2Diagnostic.__all__,
     G2Engine.__all__,
+    G2EngineFlags.__all__,
     G2Exception.__all__,
     G2Hasher.__all__,
     G2IniParams.__all__,
@@ -31,6 +32,7 @@ from .G2Config import *
 from .G2ConfigMgr import *
 from .G2Diagnostic import *
 from .G2Engine import *
+from .G2EngineFlags import *
 from .G2Exception import *
 from .G2Hasher import *
 from .G2IniParams import *
